Improve source code. I used FauxPas trial for project analysis and fixed code with analyzer recommendations. Now details in Russian: 1. Macro-based include guard. Using #pragma once 2. NSLog в релиз коде. 3. Литералы вместо констант в качестве ключей. 4. Классы без префиксов. Классы с двухбуквенными префиксами. 5. Использование self внутри init. Тут нужно разобаться и кроме того поэксплуатировать уязвимость. 6. Свойство объявлено strong, а по мнению анализатора должно быть copy. (Not fixed) 7. Неверная настройка git. Нужно добавить в gitignore *.swp *.mode1v3 *.mode2v3 *.perspectivev3 8. Hardcoded self class reference (not fixed). Нужно разобраться нужно ли нам тут это, очевидность кода уменьшает, а наследовать эти статические методы я не собираюсь. 9. Ошибка. Не использование флага -DNDEBUG (выключает assert.h). 10. Включение предупреждений.  Основное предупреждения как ошибки, все предупреждения и т.п. Перечисление есть у меня в заметках (скомпилировал и исправил случаи NSUInteger <-> int). Пришлось отключить, так как во многих системных штуках например, методе applicationLaunching есть неиспользуемые параметры (код генерируется, но до поры не используется). 11. ‘Class Prefix’ и ‘Organization’ не должны быть пустыми. Префиксы используются и сейчас, но в ручном режиме.

+++ PLTSample/UIView+PLTNestedView.m
@@ -11,10 +11,10 @@
 @implementation UIView (PLTNestedView)
 
 + (CGRect)plt_nestedViewFrame:(CGRect) containerFrame nestedScaled:(float) scale{
-  float scaleFactor = scale;
+  CGFloat scaleFactor = scale;
   
-  float xOriginDelta = containerFrame.size.width*scaleFactor / 2;
-  float yOriginDelta = containerFrame.size.height*scaleFactor /2;
+  CGFloat xOriginDelta = containerFrame.size.width*scaleFactor / 2;
+  CGFloat yOriginDelta = containerFrame.size.height*scaleFactor /2;
   
   CGPoint newOrigin = CGPointMake(containerFrame.origin.x + xOriginDelta,
                                   containerFrame.origin.y + yOriginDelta);
